Manage scrolling when the keyboard appears
This replicates the behavior of UITableViewController subclasses.
A protocol is used to declare that a conforming class should have a scrollView.
Then, a protocol extension for UIViewControllers is used in order to properly set the scrollView insets when the keyboard shows or hides.
A class extension in EditBabyViewController demonstrates the usage of this mechanism.